01-Projeto-Django/requirements.txt
django==4.2.3
Pillow==10.0.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.5.2